server/requirements.txt
Flask
flask-cors
numpy
scikit-learn
joblib
gunicorn
flask-sqlalchemy
flask-login
flask-bcrypt
flask-mail
authlib
requests
python-dotenv
